#151 feat: refatoracao da exportacao e fetch de equipamentos
